/ installx: use --remove-destination to force
cp -af still does not remove dangling symlinks, even though the manpage
seems to imply it should.  instead we use --remove-destination which
does unconditionally remove the source file beforehand, making it work
on [dangling] symbolic links.

really we should not be doing this, but until we have a stat+md5sum
version that checks the destinations for update need, we'll just always
overwrite everything and it will have the desired effect