/ mk4group: pass over links when setting perms
symlinks on linux are always 0777.  what happens, by default find does
not follow, so we're going to get mode 0777 for symlinks, but then still
give it to chmod which *will* follow the symlink.  this means we will
always set a bit in the group.  all of them will always do it

from chmod(1):

    chmod never changes the permissions of symbolic links; the chmod
    system call cannot change their permissions.  This is not a
    problem since the permissions of symbolic links are never used.
    However, for each symbolic link listed on the command line, chmod
    changes the permissions of the pointed-to file.

this assumption of chmod is false because its input does not follow the
policy chmod expects of itself when generating the file list.

not sure how this wasn't seen in testing before. seems impossible it
could pass even a simple test.