/ installx: make getopt handles spaces and count right
not sure how this got past testing, but basically it didn't work
correctly because of getopt emitting bash-style quoted strings which was
sometimes fooling the array parsing. so sometimes the quotes worked and
sometimes they didn't because it didn't have the proper context, it
needed to be in eval.

getopt would have worked the old way with -u (unquoted) if the args
didn't have spaces in them, but that should work, so it's safe.

ultimately it's a lot simpler now too.  the array was pretty useless,
just set directly from a string, that's actually what getopt intends to
give you anyways.