+ installx: add test mode, shows what would happen
this wraps command execution in a condition on test mode and prints it
otherwise.

todo: not 100% sure this survives all kinds of expansions.  it should
survive simple directory names including spaces but since it's just a
naked command execution like "${foo[@]}" as the only statement on a
line, this could really open up to all sorts of things.  but i'm not
completely certain it's bad, because theoretically the quoting tells
bash to provide all its arguments a single element of the argument
vector no matter what bytes are stored therein