/ installx: fix loop bug: erroneously used argc, not terminator
don't loop based on the arguments, this is not what we want, instead we
are supposed to loop until '--': getopt has given us reordered arguments
and placed all the options before that (and it's now a terminator), even
if the user didn't give one

this caused a bug before but don't remember what, probably when shift
was not equal to 1 because then arg count wouldn't match getopt
reordering.  or maybe because there was a '--' inserted, the count was
now off by one.  in any case this is correct usage