+ installx: support rclinks to directories
this will let us keep entire recursive dirs, not just files, in a rcfile
repo.

initial use is .fonts and .icons which we'll store in .xorg repo for
now.

UPDATE: maybe we actually won't do this come to think of it: we can just
make .fonts and .icons their own repositories.  nonetheless, the feature
is sound so might as well have it available and check it in