/ installx: only install links whose end target is a file
this bug meant that if a symlink pointed to a directory that was
executable (ie had search bits), it would have been a candidate for
install, but for exelinks, and likewise for rclinks.  of course we only
want to install symlinks if they point to files, ie scripts or config
files.

this was a bad bug, discovered when porting the script to python and
reviewing the logic, noticing the logic gap and testing it found that
indeed it would have copied ie .this -> foo/ (rclink) or this -> foo/
(exelink).

we happened not to have a scenario where it was triggered, but that was
only luck...

the xtype test for rclinks was just plain wrong, it was preferentially
looking for links that pointed to directories as one of the candidates.
this would never have been right, it seems likely that it's a
parenthesis bug:

    "$links $dots ( ( -not $exes ) -or $tdirs )"

was meant to be

    "$links $dots -not ( $exes -or $tdirs )"

but not really sure what was the thought at the time since that doesn't
make much sense either, you wouldn't want to exclude only dirs, you
would want to make sure it was pointing to a file

in any case, should be done correctly now