+ rename: new script, like rename from util-linux, w prepend/append
'rename' is built in util-linux, but on debian/ubuntu for some time it
has been compiled as 'rename.ul' and then relied on alternatives
mechanism to rename it as 'rename' in /usr/bin/.  however, in debian the
self-important maintainer of util-linux just arbitrarily imposed his
whim and removed even rename.ul so it could no longer be used even with
alternatives mechanism, despite this being the exact purpose of said
mechanism.  anyone coming from a redhat distro or having relied on the
'rename' behavior for a long time can no longer do it because of this
author's arbitrary preference for the perl implementation, which has
completely different semantics (it takes regexps and behaves
differently)

while we're at it, using empty string only works at the end for classic
'rename' so we add options '-p' and '-a' to prepend or append
(respectively) a specific string to the filenames.  '-a' works just like
"rename '' .txt files*" would, but since it only works at the end,
renames en masse with prepends would not work without doing them
manually.  having prepend capability (like to make cron jobs not glob
anymore in cron.d/) is longstanding itch with rename, now scratched.

will add usage stuff, etc later, this seems to work for now