- man: hardcode /usr/bin/man to fix systems with /bin -> /usr/bin
unfortunately it means they are the same inode on these types of
systems, so we get a duplicate entry and then count is 2, which fails.
normally would not install a system like this, but on venus we started
with the factory image, and can't really think of an easy way to fix it,
maybe a reinstall of all packages without running scripts?

not sure there are actually any systems where /usr/man won't work
anyways