- proc: psa: non-linux: make really sure not to include ourselves
tested and the old code did not always work, for example with "psa grep"
because grep was on the line by its own so it would be like "grep
[g]rep' which would match grep.

this new code goes to greater lengths to ensure it always works, by
escaping the regex and including pid string.  works when tested, but was
unable to test on other platforms it's intended to run on, but should
work.