/ proc: psf: completely reimplement, didn't do what the comment said
we don't remember the original need for this (or the reason for the name
'psf'), but we reimplemented according to the comment, which is "list
processes with a session leader matching the specified name."  the
utility of this is questionable, but we considered it a challenge,
because there's no easy way to do it in the utilities themselves.  the
fundamental problem is that negation (ie "ps -N") is an all-or-none
thing, you cannot include multiple conditions such as "match all procs
that are not session leaders named bash."

so, we had to do it ourselves in awk.  from ps output, we build separate
arrays of all procs with process id, session id, and command name, then
create an associative array indexed with session leaders matching the
requsted name, so we have an O(1) test of candidates, given a sid.
finally we go through the list again and build up an array of pids that
have the desired sid, but aren't themselves session leaders, and output
them with the separator for another 'ps' run.

the solution seems elegant, performs well, and worked first time.  its
utility remains to be seen, however... we'll see in time if we have
occasion for this