- proc: procs: 'sort -u' -> 'sort | uniq' because it runs faster
this is contrary to expectation, but running a large set of command line
history through them both results in 'sort | uniq' being slightly
faster:

   $ time -p cat var/log/cmdhist* | sort -u >/dev/null
  real 6.13
  user 5.93
  sys 0.21

   $ time -p cat var/log/cmdhist* | sort | uniq >/dev/null
  real 5.23
  user 5.15
  sys 0.28

unclear why this is.  at first assumed multicore, because the 'sort' and
the 'uniq' can run on different processors, but actually 'sort' cannot
output anything until it has completed the sort, so 'uniq', even if
running on a different cpu, cannot run in parallel.  so 'sort -u'
probably should be faster since it can do deduplication at the same time
as sorting without any separate IO.

yet, it's true...