- proc: psl: new name for psf, which we will use for ps forest instead
this session-leader-name matching command should not be 'psf', which
makes more sense for a command that displays a lineage tree rooted at
given pid, ie showing the "forest" from a root pid (ps-"f"orest).  since
we want to implement that one, and think that command will be used more,
we'll rename this more useless one 'psl' with the 'l' as in 'l'eader.