+ proc: psf: re-add a 'psf' that lists full tree of descendant procs
'f' as in 'forest', ie, show the whole hierarchical tree rooted at the
given pids

we really need awk to do this, because invoking 'ps' repeatedly on the
list of child procs and recursing would (1) have a lot of overhead, and
(2) not be correct, because it would not work on a single snapshot.  we
only really need the info once from 'ps' and can do our own filtering,
sorting, list creation, then give it back to 'ps' again for the final
display of selected pids.

any processes that no longer exist will be silently ignored