+ proc: if '-p' given as first arg, display pids only
if user only wants pids to be output (for postprocessing by another
program), they will give the '-p' option

note that this will still be tabular for the functions that output with
mktable(), and if the user wanted to reformat, the obvious "fmt -1"
still preserves lhs-space before words in the column, which will make it
sort differently.  so we should probably work in the future towards
making '-p' display one column only (left justified), and optional
comma-delim.  this will require switching to getopt(1).  for now it's
enough to give the arg and have pids come out instead of comm= in
mktable case of psflags in other cases