+ proc: noself: allow callers to set exclk=1 and kthreads will be ignored
we usually do not want to match kthreads; we have an upcoming commit
where we will list all terminals without a controlling tty, and this
includes all kthreads, which we will not want, so we will have to remove
those.  for existing users of ps_noself_select() it would not have
returned kthreads anyways, so it was not necessary to filter.  however
it's harmless if they are filtered, and this new list-by-tty function
will otherwise be the same as other users of ps_noself_select(), so we
will decide on its presence via environment, so caller can select.

fortunately this is simple to do because kthreadd is pid2 and any of its
children are kernel threads, so all kernel threads have ppid==2

todo: what about indirect kthreads? have not seen this in practice,
let's see if it ever shows up and worry about it then.  session 0 might
work, all kernel threads seem to have sid 0 in 'ps'.  but pgrep treats
session 0 as the session of the program, so want to keep using pid==2
for now and see if this actually comes up in the real world