+ exes: new invokes: bins, pyscripts, shscripts, scripts (select by magic)
use file magic to determine what kind of file, so user can select only
python, only shell, all scripts, just [compiled] binaries, or all
executables (as before).

as a side effect of the new looping structure, it will also now process
multiple arguments in a more sane way, previously it would just print
the basename even if a full path was given.  now it prints absolute path
if the argument contains a slash, but otherwise prints a relative path
(ie, the common use case, for files in the current directory)