+ cmpfuncs: new script, compares same-named funcs in two source files
this came up working on libshared from GothenburgBitFactory,
timewarrior's DatetimeParser.cpp shadows several private functions in
libshared's Datetime.cpp and we want to change some of them in
libshared.  since they are duplicated we need a script to easily compare
so we can track the changes and re-implement them in timewarrior.

this is a general enough need that we made the script to just take two
source files, and optionally function names following that, so it can be
used on any project, rather than hardcodeing the filenames for the
libshared case.

it ignores the signature line (by duplicating the one from the second
file as the first line given to diff, even for the first file) because
otherwise there would be false diffs due to the class name being
different, when the function code itself is otherwise the same.