/ daemonize: close stdout and stderr of logger itself
while logger is set to absorb output and stderr of the process
that daemonizes, it does not itself need to output anywhere... if we
don't close stdout and stderr or set them to /dev/null then they will
leave stray descriptors open on the invoking tty.  for something
supposed to run in the background this is no good.