/ exes: print full/relative path root on *any* non-cwd argument
we were testing whether '/' was present before but really we want to
print the initial path string on *any* argument that isn't the current
directory (ie the defaulted '.' when no args are given).  not sure this
covers all cases with correct behavior but it covers the most common use
cases, we'll see if others come up in practice.