/ blkfree: handle systems without modern findmnt by scraping df
older systems will not have the flags we need from findmnt, either
bytes-only numbers or can print available disk space.  so we have to
scrape 'df' output.  amazed that there is no other way.  it looks like
we can do "stat -c --file-system %a" but we cannot print mount point
that way either, the %m format does not seem to be working for me so
we'll have to scrape 'df' output.  this is just an older system anyways
and will probably be gone soon