+ exes: remove restriction on args being dirs, so can pass file list
passing a list of files we came up with by other means than ourselves
allows us to use this program as a kind of filter on all the files in
the argument vector, rather than just all files in directories named in
the argument vector.

at first we tried to change logic around to treat files differently, but
it became complicated and would not have been easy in a short amount of
time.  but then realized it will just work as-is as long as we're ok
with using 'find' once for each file, we already have maxdepth 1 so this
should work without any changes, just remove the restriction on
arguments having to be directores.  voila...