/ lsa: return the status code from 'ls', unless pipeline had bad exit
normally we would want the return code from the list command itself, ie
"lsa -d /lkjalskdjf" should return false, since the file is not present.
we do check the return code of the pipeline though and give a bad exit
from that if it had one, because the user ought to know about that and
not just blindly use the 'ls' return code, the results might make no
sense

it occurs to me now, we could also use "set -o pipefail" maybe, it would
make things easier, this is something to think about for future.  for
now we'll leave it like this since it was already implemented and should
work fine.  it also leaves options open in case this file gets imported
rather than executed, because otherwise it would have to save the
pipefail setting in each function, run the code, and then restore it,
which would be a pain