/ memstrings: strings: fix bug where "strings -U" would segfault
not clear why, and it probably shouldn't, but it seems to segfault when
we try to decode it, even if we use hex bytes.  it really shouldn't
because root runs strings a lot so this would be easy to get a hack
vector.  strings has had several security bugs in the past already,
harrumph.  don't like this but it's still useful and anyways it works as
non-root.  this fixes the issue, to use the native encoding looks like.
or, at least it doesn't appear with a few tests we ran.